requirements.txt
requests
mastercard-oauth1-signer